Remove singleton to fix issues when chain multiples cmd run
We have some issue when running acceptance test, more generally
when we use to execute scan cmd multiples times.
We were using global singletons for provider and resources suppliers
managment which lead us to improper state in the second scan run.
We should avoid this in the future and make proper initialization of our
dependencies maybe using a dependency injection container.